move `rgbToHex` to color utils
